+++ NEWS.md
@@ -197,11 +197,16 @@
     correctly, rather than `nil` as in previous releases.  It's also
     considerably faster now that it doesn't use `pcall` any more.
 
-  - `std.functional.bind` returns functions that propagate trailing `nil`
-    arguments correctly.
+  - `std.functional.any`, `std.functional.bind` and
+    `std.functional.compose` return functions that propagate trailing
+    `nil` arguments correctly.
 
   - `std.functional.memoize` now considers trailing nil arguments when
     looking up memoized value for those particular arguments.
+
+  - `std.functional.filter`, `std.functional.map` and
+    `std.functional.reduce` now pass trailing nil arguments to their
+    iterator function correctly.
 
   - You can now derive other types from `std.set` by passing a `_type`
     field in the init argument, just like the other table argument